pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "liq-sweep"
version = "0.1.0"
description = "Realtime NSE intraday breakout scanner and execution research workstation."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "numpy>=1.26",
  "polars>=1.0",
  "prometheus-client>=0.20",
  "pydantic>=2.7",
  "PyYAML>=6.0",
  "structlog>=24.1",
  "tenacity>=8.3",
  "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
  "hypothesis>=6.100",
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "respx>=0.21",
]

[project.scripts]
liq-sweep-scanner = "src.app:main"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-ra"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.12"
strict = true
